requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
seaborn
pandas
opencv-python
pillow
streamlit